+++ src/main/java/io/dataspaceconnector/model/truststore/TruststoreDesc.java
@@ -12,6 +12,10 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ *  Contributors:
+ *       sovity GmbH
+ *
  */
 package io.dataspaceconnector.model.truststore;
 
@@ -41,9 +45,4 @@ public class TruststoreDesc extends Description {
      * The password of the trust store.
      */
     private String password;
-
-    /**
-     * Alias for the key store.
-     */
-    private String alias;
 }